Expose rotation degrees from video metadata.
In some platforms, like Android, the value of the width and height
is not representative of the actual orientation of the video.
To accurately detect if the video is in landscape or portrait mode
we need to also check the value of the degrees.
This patch exposes a new property for the onLoad event inside the
naturalSize for the rotationDegrees. It also modifies the value of
the orientation to portrait if the degrees are equal to 90 or 270.